feat: move data validation and conversion logic to helper functions and update trailer fields for admin form and dashboard types
