+++ src/types/dashboardSetTypes.ts
@@ -11,7 +11,8 @@ export type SetAdminEvent = {
     discord_url: string | null,
     twitch_url: string | null,
     presskit_url: string | null,
-    trailer_youtube_id: string | null,
+    trailer_url: string | null,
+    trailer_poster_url: string | null,
     description_headline: string,
     description: string,
     schedule_visible_from: string | null,
